backend/tests/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-flask>=1.2.0
coverage>=7.0.0
factory-boy>=3.2.0
faker>=18.0.0
responses>=0.23.0
freezegun>=1.2.0